+++ src/pyuvdata/__init__.py
@@ -48,7 +48,14 @@ warnings.filterwarnings("ignore", message="numpy.ufunc size changed")
 
 from .telescopes import Telescope  # noqa
 from .telescopes import get_telescope  # noqa  # NB: get_telescopes is deprecated
-from .uvbeam import UVBeam  # noqa
+from .uvbeam import (  # noqa
+    AiryBeam,
+    BeamInterface,
+    GaussianBeam,
+    ShortDipoleBeam,
+    UniformBeam,
+    UVBeam,
+)
 from .uvcal import UVCal  # noqa
 from .uvdata import FastUVH5Meta  # noqa
 from .uvdata import UVData  # noqa
@@ -60,6 +67,11 @@ __all__ = [
     "UVCal",
     "UVFlag",
     "UVBeam",
+    "BeamInterface",
+    "AiryBeam",
+    "GaussianBeam",
+    "ShortDipoleBeam",
+    "UniformBeam",
     "Telescope",
     "get_telescope",
 ]